refactor(settings): split render sections into per-section files
Finish the settings module split started in #185/PR #213. The 652-line
src/settings.ts god class is now a 15-line facade; each UI section
lives in its own file under src/settings/ and McpSettingsTab is a
thin composer.

New files under src/settings/:
- tab.ts                  — McpSettingsTab composer (38 lines)
- server-section.ts       — Server Status + Server Settings (241 lines)
- https-section.ts        — regenerate/custom-TLS branch (60 lines)
- tls-section.ts          — custom TLS picker + error renderer (133 lines)
- mcp-config-section.ts   — client JSON fragment + copy (57 lines)
- modules-section.ts      — per-module + per-tool toggles (103 lines)
- allowlist-section.ts    — executeCommand allowlist textarea (35 lines)
- diagnostics-section.ts  — debug, log file, copy debug, clear (60 lines)
- validation.ts           — createValidationError + isValidIPv4 + key gen (52 lines)
- migrations.ts           — already extracted in PR #213

Every file under src/settings/ is now ≤ 250 lines (the largest is
server-section.ts at 241). src/settings.ts re-exports McpSettingsTab,
migrateSettings, generateAccessKey, isValidIPv4, and
createValidationError so existing import paths in main.ts, tests, and
docs keep working unchanged.

Pure code move — zero UI changes. Verified end-to-end by driving
Obsidian via Xvfb + CDP and diffing before/after screenshots of the
settings tab (attached to the PR).

Closes #214
