+++ src/settings.ts
@@ -1,652 +1,15 @@
-import { App, Notice, PluginSettingTab, Setting } from 'obsidian';
-import { randomBytes } from 'crypto';
-import type McpPlugin from './main';
-import type { ModuleRegistration } from './registry/types';
-import { t, type TranslationKey } from './lang/helpers';
-import { clearLogFile, getLogFilePath } from './utils/log-file';
-import { DebugInfoModal } from './ui/debug-info-modal';
-import {
-  CustomTlsError,
-  loadAndValidateCustomTls,
-} from './server/custom-tls';
-import { pickFile } from './utils/file-picker';
-import { reportError } from './utils/report-error';
-
-export class McpSettingsTab extends PluginSettingTab {
-  plugin: McpPlugin;
-
-  constructor(app: App, plugin: McpPlugin) {
-    super(app, plugin);
-    this.plugin = plugin;
-  }
-
-  display(): void {
-    const { containerEl } = this;
-    containerEl.empty();
-
-    this.renderServerStatus(containerEl);
-    this.renderServerSettings(containerEl);
-    this.renderMcpConfig(containerEl);
-    this.renderModuleToggles(containerEl);
-    this.renderExecuteCommandAllowlist(containerEl);
-    this.renderDiagnostics(containerEl);
-  }
-
-  private renderExecuteCommandAllowlist(containerEl: HTMLElement): void {
-    containerEl.createEl('h2', { text: 'Execute Command Allowlist' });
-
-    containerEl.createEl('p', {
-      cls: 'setting-item-description',
-      text: 'The plugin_execute_command tool refuses every call by default. To enable specific Obsidian commands for MCP execution, list their ids here (one per line). Leave empty to keep the tool disabled.',
-    });
-
-    const current = this.plugin.settings.executeCommandAllowlist.join('\n');
-    const textarea = containerEl.createEl('textarea', {
-      cls: 'mcp-execute-command-allowlist',
-      attr: { rows: '6', placeholder: 'app:reload\neditor:save-file' },
-    });
-    textarea.value = current;
-    textarea.addEventListener('change', () => {
-      const next = textarea.value
-        .split(/\r?\n/)
-        .map((line) => line.trim())
-        .filter((line) => line.length > 0);
-      this.plugin.settings.executeCommandAllowlist = next;
-      void this.plugin.saveSettings();
-    });
-  }
-
-  private scheme(): 'http' | 'https' {
-    return this.plugin.settings.httpsEnabled ? 'https' : 'http';
-  }
-
-  private renderServerStatus(containerEl: HTMLElement): void {
-    containerEl.createEl('h2', { text: t('heading_server_status') });
-
-    const isRunning = this.plugin.httpServer?.isRunning ?? false;
-    const port = this.plugin.settings.port;
-    const clients = this.plugin.httpServer?.connectedClients ?? 0;
-
-    const address = this.plugin.settings.serverAddress;
-    const url = `${this.scheme()}://${address}:${String(port)}`;
-    const statusText = isRunning
-      ? clients === 1
-        ? t('status_running_one', { url })
-        : t('status_running_many', { url, count: clients })
-      : t('status_stopped');
-
-    const setting = new Setting(containerEl)
-      .setName(t('setting_status_name'))
-      .setDesc(statusText);
-
-    if (isRunning) {
-      setting.addExtraButton((btn) =>
-        btn
-          .setIcon('refresh-cw')
-          .setTooltip(t('tooltip_restart_server'))
-          .onClick(() => {
-            this.plugin
-              .restartServer()
-              .then(() => {
-                this.display();
-              })
-              .catch(reportError('restart server', this.plugin.logger));
-          }),
-      );
-    }
-
-    setting.addToggle((toggle) =>
-      toggle
-        .setValue(isRunning)
-        .setTooltip(isRunning ? t('tooltip_stop_server') : t('tooltip_start_server'))
-        .onChange((value) => {
-          const action = value
-            ? this.plugin.startServer()
-            : this.plugin.stopServer();
-          action
-            .then(() => {
-              this.display();
-            })
-            .catch(
-              reportError(
-                value ? 'start server' : 'stop server',
-                this.plugin.logger,
-              ),
-            );
-        }),
-    );
-  }
-
-  private renderServerSettings(containerEl: HTMLElement): void {
-    containerEl.createEl('h2', { text: t('heading_server_settings') });
-
-    const addressSetting = new Setting(containerEl)
-      .setName(t('setting_server_address_name'))
-      .setDesc(t('setting_server_address_desc'));
-    const addressError = createValidationError(addressSetting);
-    addressSetting.addText((text) =>
-      text
-        .setPlaceholder('127.0.0.1')
-        .setValue(this.plugin.settings.serverAddress)
-        .onChange(async (value) => {
-          if (isValidIPv4(value)) {
-            addressError.clear();
-            this.plugin.settings.serverAddress = value;
-            await this.plugin.saveSettings();
-          } else {
-            addressError.show(t('error_invalid_ipv4'));
-          }
-        }),
-    );
-
-    if (this.plugin.settings.serverAddress !== '127.0.0.1') {
-      addressSetting.descEl.createEl('br');
-      addressSetting.descEl.createEl('strong', {
-        text: t('warning_non_localhost'),
-      });
-    }
-
-    const portSetting = new Setting(containerEl)
-      .setName(t('setting_port_name'))
-      .setDesc(t('setting_port_desc'));
-    const portError = createValidationError(portSetting);
-    const portStartError = createValidationError(portSetting);
-    const startFailure = this.plugin.lastStartError;
-    if (startFailure && startFailure.port === this.plugin.settings.port) {
-      portStartError.show(
-        t('settings_port_in_use_error', { port: startFailure.port }),
-      );
-    }
-    portSetting.addText((text) =>
-      text
-        .setPlaceholder('28741')
-        .setValue(String(this.plugin.settings.port))
-        .onChange(async (value) => {
-          const port = parseInt(value, 10);
-          if (/^\d+$/.test(value) && !isNaN(port) && port > 0 && port < 65536) {
-            portError.clear();
-            portStartError.clear();
-            this.plugin.settings.port = port;
-            await this.plugin.saveSettings();
-          } else {
-            portError.show(t('error_invalid_port'));
-          }
-        }),
-    );
-
-    const serverUrl = `${this.scheme()}://${this.plugin.settings.serverAddress}:${String(this.plugin.settings.port)}/mcp`;
-    new Setting(containerEl)
-      .setName(t('setting_server_url_name'))
-      .setDesc(serverUrl)
-      .addExtraButton((btn) =>
-        btn
-          .setIcon('copy')
-          .setTooltip(t('tooltip_copy_server_url'))
-          .onClick(() => {
-            navigator.clipboard
-              .writeText(serverUrl)
-              .then(() => {
-                new Notice(t('notice_server_url_copied'));
-              })
-              .catch(reportError('copy server URL', this.plugin.logger));
-          }),
-      );
-
-    new Setting(containerEl)
-      .setName(t('setting_auth_enabled_name'))
-      .setDesc(t('setting_auth_enabled_desc'))
-      .addToggle((toggle) =>
-        toggle
-          .setValue(this.plugin.settings.authEnabled)
-          .onChange(async (value) => {
-            this.plugin.settings.authEnabled = value;
-            await this.plugin.saveSettings();
-            this.display();
-          }),
-      );
-
-    if (this.plugin.settings.authEnabled) {
-      new Setting(containerEl)
-        .setName(t('setting_access_key_name'))
-        .setDesc(t('setting_access_key_desc'))
-        .addText((text) =>
-          text
-            .setPlaceholder(t('placeholder_access_key'))
-            .setValue(this.plugin.settings.accessKey)
-            .onChange(async (value) => {
-              this.plugin.settings.accessKey = value;
-              await this.plugin.saveSettings();
-            }),
-        )
-        .addExtraButton((btn) =>
-          btn
-            .setIcon('copy')
-            .setTooltip(t('tooltip_copy_access_key'))
-            .onClick(() => {
-              navigator.clipboard
-                .writeText(this.plugin.settings.accessKey)
-                .then(() => {
-                  new Notice(t('notice_access_key_copied'));
-                })
-                .catch(reportError('copy access key', this.plugin.logger));
-            }),
-        )
-        .addExtraButton((btn) =>
-          btn
-            .setIcon('refresh-cw')
-            .setTooltip(t('tooltip_generate'))
-            .onClick(() => {
-              this.plugin.settings.accessKey = generateAccessKey();
-              this.plugin
-                .saveSettings()
-                .then(() => {
-                  this.display();
-                })
-                .catch(reportError('save settings', this.plugin.logger));
-            }),
-        );
-    }
-
-    new Setting(containerEl)
-      .setName(t('setting_https_name'))
-      .setDesc(t('setting_https_desc'))
-      .addToggle((toggle) =>
-        toggle
-          .setValue(this.plugin.settings.httpsEnabled)
-          .onChange(async (value) => {
-            this.plugin.settings.httpsEnabled = value;
-            await this.plugin.saveSettings();
-            this.display();
-          }),
-      );
-
-    if (this.plugin.settings.httpsEnabled) {
-      if (!this.plugin.settings.useCustomTls) {
-        const hasCert = this.plugin.settings.tlsCertificate !== null;
-        new Setting(containerEl)
-          .setName(t('setting_tls_cert_name'))
-          .setDesc(
-            hasCert
-              ? t('setting_tls_cert_desc_present')
-              : t('setting_tls_cert_desc_absent'),
-          )
-          .addExtraButton((btn) =>
-            btn
-              .setIcon('refresh-cw')
-              .setTooltip(t('tooltip_regenerate_cert'))
-              .onClick(() => {
-                this.plugin
-                  .regenerateTlsCertificate()
-                  .then(() => {
-                    new Notice(t('notice_tls_regenerated'));
-                    this.display();
-                  })
-                  .catch(
-                    reportError(
-                      'regenerate TLS certificate',
-                      this.plugin.logger,
-                    ),
-                  );
-              }),
-          );
-      }
-
-      new Setting(containerEl)
-        .setName(t('setting_custom_tls_toggle_name'))
-        .setDesc(t('setting_custom_tls_toggle_desc'))
-        .addToggle((toggle) =>
-          toggle
-            .setValue(this.plugin.settings.useCustomTls)
-            .onChange(async (value) => {
-              this.plugin.settings.useCustomTls = value;
-              await this.plugin.saveSettings();
-              this.display();
-            }),
-        );
-
-      if (this.plugin.settings.useCustomTls) {
-        this.renderCustomTlsGroup(containerEl);
-      }
-    }
-
-    new Setting(containerEl)
-      .setName(t('setting_autostart_name'))
-      .setDesc(t('setting_autostart_desc'))
-      .addToggle((toggle) =>
-        toggle
-          .setValue(this.plugin.settings.autoStart)
-          .onChange(async (value) => {
-            this.plugin.settings.autoStart = value;
-            await this.plugin.saveSettings();
-          }),
-      );
-  }
-
-  private renderCustomTlsGroup(containerEl: HTMLElement): void {
-    containerEl.createEl('h3', { text: t('heading_custom_tls_group') });
-
-    const certSetting = new Setting(containerEl)
-      .setName(t('setting_custom_tls_cert_name'))
-      .setDesc(
-        t('setting_custom_tls_cert_desc', {
-          path:
-            this.plugin.settings.customTlsCertPath ?? t('label_no_file_selected'),
-        }),
-      );
-    const certError = createValidationError(certSetting);
-    certSetting.addButton((btn) =>
-      btn.setButtonText(t('button_browse')).onClick(() => {
-        this.pickCustomTlsPath('cert').catch(
-          reportError('pick custom TLS certificate', this.plugin.logger),
-        );
-      }),
-    );
-
-    const keySetting = new Setting(containerEl)
-      .setName(t('setting_custom_tls_key_name'))
-      .setDesc(
-        t('setting_custom_tls_key_desc', {
-          path:
-            this.plugin.settings.customTlsKeyPath ?? t('label_no_file_selected'),
-        }),
-      );
-    const keyError = createValidationError(keySetting);
-    keySetting.addButton((btn) =>
-      btn.setButtonText(t('button_browse')).onClick(() => {
-        this.pickCustomTlsPath('key').catch(
-          reportError('pick custom TLS key', this.plugin.logger),
-        );
-      }),
-    );
-
-    const { customTlsCertPath, customTlsKeyPath } = this.plugin.settings;
-    if (customTlsCertPath && customTlsKeyPath) {
-      void loadAndValidateCustomTls(customTlsCertPath, customTlsKeyPath).then(
-        () => {
-          certError.clear();
-          keyError.clear();
-        },
-        (err: unknown) => {
-          this.showCustomTlsError(err, certError, keyError);
-        },
-      );
-    }
-  }
-
-  private async pickCustomTlsPath(kind: 'cert' | 'key'): Promise<void> {
-    const title =
-      kind === 'cert' ? t('dialog_title_pick_cert') : t('dialog_title_pick_key');
-    const filters =
-      kind === 'cert'
-        ? [
-            { name: 'PEM certificate', extensions: ['pem', 'crt', 'cer'] },
-            { name: 'All files', extensions: ['*'] },
-          ]
-        : [
-            { name: 'PEM private key', extensions: ['pem', 'key'] },
-            { name: 'All files', extensions: ['*'] },
-          ];
-    const chosen = await pickFile({ title, filters });
-    if (!chosen) return;
-
-    if (kind === 'cert') {
-      this.plugin.settings.customTlsCertPath = chosen;
-    } else {
-      this.plugin.settings.customTlsKeyPath = chosen;
-    }
-    await this.plugin.saveSettings();
-    this.display();
-  }
-
-  private showCustomTlsError(
-    err: unknown,
-    certError: ValidationErrorController,
-    keyError: ValidationErrorController,
-  ): void {
-    if (!(err instanceof CustomTlsError)) {
-      certError.show(err instanceof Error ? err.message : String(err));
-      keyError.clear();
-      return;
-    }
-    const key: TranslationKey = `error_custom_tls_${err.code}`;
-    const message = t(key);
-    switch (err.code) {
-      case 'cert_not_readable':
-      case 'invalid_cert':
-      case 'cert_expired':
-        certError.show(message);
-        keyError.clear();
-        return;
-      case 'key_not_readable':
-      case 'invalid_key':
-        keyError.show(message);
-        certError.clear();
-        return;
-      case 'key_cert_mismatch':
-        certError.show(message);
-        keyError.show(message);
-        return;
-    }
-  }
-
-  private renderMcpConfig(containerEl: HTMLElement): void {
-    containerEl.createEl('h2', { text: t('heading_mcp_client_config') });
-
-    new Setting(containerEl)
-      .setName(t('setting_client_config_name'))
-      .setDesc(t('setting_client_config_desc'))
-      .addExtraButton((btn) =>
-        btn
-          .setIcon('copy')
-          .setTooltip(t('tooltip_copy_config'))
-          .onClick(() => {
-            navigator.clipboard
-              .writeText(this.buildMcpConfigJson())
-              .then(() => {
-                new Notice(t('notice_config_copied'));
-              })
-              .catch(reportError('copy MCP client config', this.plugin.logger));
-          }),
-      );
-  }
-
-  private buildMcpConfigJson(): string {
-    const address = this.plugin.settings.serverAddress;
-    const port = this.plugin.settings.port;
-    const accessKey = this.plugin.settings.accessKey;
-    const authEnabled = this.plugin.settings.authEnabled;
-    const url = `${this.scheme()}://${address}:${String(port)}/mcp`;
-
-    const config: Record<string, unknown> = { url };
-
-    if (authEnabled && accessKey) {
-      config.headers = {
-        Authorization: `Bearer ${accessKey}`,
-      };
-    }
-
-    const full = JSON.stringify({ obsidian: config }, null, 2);
-    const lines = full.split('\n');
-    return lines
-      .slice(1, -1)
-      .map((line) => line.slice(2))
-      .join('\n');
-  }
-
-  private renderDiagnostics(containerEl: HTMLElement): void {
-    containerEl.createEl('h2', { text: t('heading_diagnostics') });
-
-    new Setting(containerEl)
-      .setName(t('setting_debug_name'))
-      .setDesc(t('setting_debug_desc'))
-      .addToggle((toggle) =>
-        toggle
-          .setValue(this.plugin.settings.debugMode)
-          .onChange(async (value) => {
-            this.plugin.settings.debugMode = value;
-            this.plugin.logger.updateOptions({ debugMode: value });
-            await this.plugin.saveSettings();
-          }),
-      );
-
-    new Setting(containerEl)
-      .setName(t('setting_log_file_name'))
-      .setDesc(getLogFilePath(this.plugin));
-
-    new Setting(containerEl)
-      .setName(t('setting_copy_debug_info_name'))
-      .setDesc(t('setting_copy_debug_info_desc'))
-      .addExtraButton((btn) =>
-        btn
-          .setIcon('copy')
-          .setTooltip(t('tooltip_copy_debug_info'))
-          .onClick(() => {
-            new DebugInfoModal(this.plugin.app, this.plugin).open();
-          }),
-      );
-
-    new Setting(containerEl)
-      .setName(t('setting_clear_log_name'))
-      .setDesc(t('setting_clear_log_desc'))
-      .addExtraButton((btn) =>
-        btn
-          .setIcon('trash')
-          .setTooltip(t('tooltip_clear_log'))
-          .onClick(() => {
-            clearLogFile(this.plugin)
-              .then(() => {
-                new Notice(t('notice_log_cleared'));
-              })
-              .catch(reportError('clear log file', this.plugin.logger));
-          }),
-      );
-  }
-
-  private renderModuleToggles(containerEl: HTMLElement): void {
-    const modules = this.plugin.registry.getModules();
-    const coreModules = modules.filter((r) => !r.module.metadata.group);
-    const extrasModules = modules.filter(
-      (r) => r.module.metadata.group === 'extras',
-    );
-
-    containerEl.createEl('h2', { text: t('heading_feature_modules') });
-
-    if (modules.length === 0) {
-      containerEl.createEl('p', {
-        text: t('message_no_modules'),
-        cls: 'setting-item-description',
-      });
-    }
-
-    for (const registration of coreModules) {
-      this.renderModuleRow(containerEl, registration);
-    }
-
-    if (extrasModules.length > 0) {
-      containerEl.createEl('h2', { text: t('heading_extras') });
-      for (const registration of extrasModules) {
-        this.renderExtrasToolRows(containerEl, registration);
-      }
-    }
-
-    new Setting(containerEl).addButton((btn) =>
-      btn.setButtonText(t('button_refresh_modules')).onClick(() => {
-        this.plugin.refreshModules();
-        this.display();
-      }),
-    );
-  }
-
-  private renderExtrasToolRows(
-    containerEl: HTMLElement,
-    registration: ModuleRegistration,
-  ): void {
-    const moduleId = registration.module.metadata.id;
-    const tools = registration.module.tools();
-
-    for (const tool of tools) {
-      new Setting(containerEl)
-        .setName(tool.name)
-        .setDesc(tool.description)
-        .addToggle((toggle) =>
-          toggle
-            .setValue(registration.toolStates[tool.name] ?? false)
-            .onChange(async (value) => {
-              this.plugin.registry.setToolEnabled(moduleId, tool.name, value);
-              this.plugin.settings.moduleStates = this.plugin.registry.getState();
-              await this.plugin.saveSettings();
-            }),
-        );
-    }
-  }
-
-  private renderModuleRow(
-    containerEl: HTMLElement,
-    registration: ModuleRegistration,
-  ): void {
-    const { metadata } = registration.module;
-
-    const card = containerEl.createDiv({ cls: 'mcp-module-card' });
-
-    new Setting(card)
-      .setName(metadata.name)
-      .setDesc(metadata.description)
-      .setClass('mcp-module-card-header')
-      .addToggle((toggle) =>
-        toggle.setValue(registration.enabled).onChange(async (value) => {
-          if (value) {
-            this.plugin.registry.enableModule(metadata.id);
-          } else {
-            this.plugin.registry.disableModule(metadata.id);
-          }
-          this.plugin.settings.moduleStates = this.plugin.registry.getState();
-          await this.plugin.saveSettings();
-        }),
-      );
-  }
-}
-
-interface ValidationErrorController {
-  show: (message: string) => void;
-  clear: () => void;
-}
-
-function createValidationError(setting: Setting): ValidationErrorController {
-  let errorEl: HTMLElement | null = null;
-  return {
-    show: (message: string): void => {
-      if (errorEl) {
-        errorEl.textContent = message;
-        return;
-      }
-      errorEl = setting.descEl.createEl('div', {
-        cls: 'mcp-settings-error',
-        text: message,
-      });
-    },
-    clear: (): void => {
-      if (errorEl) {
-        errorEl.remove();
-        errorEl = null;
-      }
-    },
-  };
-}
-
-export function generateAccessKey(): string {
-  return randomBytes(32).toString('hex');
-}
-
-export function isValidIPv4(value: string): boolean {
-  const parts = value.split('.');
-  if (parts.length !== 4) return false;
-  return parts.every((part) => {
-    const num = Number(part);
-    return /^\d{1,3}$/.test(part) && num >= 0 && num <= 255;
-  });
-}
-
-// Settings migration logic lives in src/settings/migrations.ts — extracted
-// per #185 so each per-version hop is unit-testable without pulling in the
-// Obsidian Plugin lifecycle.
+/**
+ * Public facade for the settings module. The actual implementation lives
+ * under `src/settings/` (migrations, per-section renderers, the composer
+ * tab, and shared validation helpers). Everything here is re-exported so
+ * existing import paths (`import { McpSettingsTab, migrateSettings } from
+ * './settings'`) keep working unchanged.
+ */
+export { McpSettingsTab } from './settings/tab';
 export { migrateSettings } from './settings/migrations';
+export {
+  generateAccessKey,
+  isValidIPv4,
+  createValidationError,
+  type ValidationErrorController,
+} from './settings/validation';
